add index function
