+++ config/routes.rb
@@ -2,9 +2,7 @@ Rails.application.routes.draw do
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
   namespace :api do
     namespace :v1 do
-      namespace :articles do
-        post "/", to: "article#new"
-      end
+      resources :articles
     end
   end
 end
