+++ test/rest/article.http
@@ -1,3 +1,7 @@
+GET http://localhost:3000/api/v1/articles
+
+
+###
 POST http://localhost:3000/api/v1/articles
 content-type: application/json
